requirements.txt
flask==3.0.0
pypdf==3.17.0
pdf2image==1.16.3
pillow==10.0.0
gunicorn==21.2.0
pyngrok==7.0.0
streamlit
streamlit-pdf-viewer
fastapi
uvicorn
python-multipart
